Sema: Move some code from CSLookahead.cpp to Subtyping.cpp
